exceptions: add try_with_catch_all, use everywhere
Add function that implements "try { <Try> } catch (...) { <Catch> }"
pattern and is to be used in the library instead of using try...catch
directly. This will allow us to build the code with exceptions disabled
by modifying this function instead of modifying multiple parts of the
code.

Use try_with_catch_all in the existing code.