exceptions: add throw_logic_error, use everywhere
Add function that simply throws std::logic_error with specified message
and use it everywhere in the code instead of using "throw" directly.
This will allow us to build the code with exceptions disabled by
modifying only this function.

The function is to be used for reporting assertion-like errors, when
some API is used incorrectly (it always signals a bug in the code that
uses the API).