Separate nosync:ux from "core" nosync
Move code and interfaces which are either UNIX-specific (e.g. use file
descriptors) or dependent on external state (files, networking,
threads) from "nosync" to "nosync::ux" namespace. Leave aliases in
headers with the old names for now.

The aliases are considered deprecated API and will be removed in the
future.