synchronized_queue: add pop_group/try_pop_group
Add batch pop operations to synchronized_queue:
 - pop_group - blocking call that waits for at least one element in the
queue but returns more elements if available (up to specified limit),
 - try_pop_group - like pop_group but blocks with timeout and returns
empty group iff timeout is exceeded.

Batch operations are more efficient when many elements are pulled from
a queue.

We'll need a variant with timeout to be able to use synchronized_queue
as event source with new type of event loop.